# Partially fix whitespace parsing performance regression