@Gabriel439 2019-11-04
This is about as fast as I can get with simple changes.  Getting fully back to the original performance would require much more invasive fixes, but I think this is probably good enough.